+++ modules/firebase/role.ts
@@ -5,7 +5,7 @@
 //
 // SPDX-License-Identifier: MIT
 //
-import { type UserType } from '@/modules/firebase/utils'
+import { type UserType } from '@stanfordbdhg/engagehf-models'
 import { upperFirst } from '@/packages/design-system/src/utils/misc'
 
 export const stringifyType = (type: UserType) => upperFirst(type)
