+++ modules/user/queries.tsx
@@ -5,19 +5,18 @@
 //
 // SPDX-License-Identifier: MIT
 //
+import { UserType } from '@stanfordbdhg/engagehf-models'
 import { queryOptions } from '@tanstack/react-query'
 import { query, where } from 'firebase/firestore'
 import { queryClient } from '@/app/ReactQueryClientProvider'
 import { docRefs, getCurrentUser, refs } from '@/modules/firebase/guards'
+import { type Invitation, type Organization } from '@/modules/firebase/models'
 import { mapAuthData } from '@/modules/firebase/user'
 import {
   getDocData,
   getDocDataOrThrow,
   getDocsData,
-  type Invitation,
-  type Organization,
   type UserAuthenticationInformation,
-  UserType,
 } from '@/modules/firebase/utils'
 
 export const getNonAdminInvitations = (organizationIds: string[]) =>
