+++ routes/~_dashboard/~patients/~$id/Allergies.tsx
@@ -9,7 +9,7 @@ import { useRouter } from '@tanstack/react-router'
 import { createColumnHelper } from '@tanstack/table-core'
 import { Plus } from 'lucide-react'
 import { useMemo } from 'react'
-import { stringifyAllergyType } from '@/modules/firebase/models/allergy'
+import { stringifyAllergyType } from '@/modules/firebase/allergy'
 import { Button } from '@/packages/design-system/src/components/Button'
 import { DataTable } from '@/packages/design-system/src/components/DataTable'
 import { useOpenState } from '@/packages/design-system/src/utils/useOpenState'
