+++ routes/~_dashboard/~users/~$id.tsx
@@ -6,11 +6,12 @@
 // SPDX-License-Identifier: MIT
 //
 import { updateDoc } from '@firebase/firestore'
+import { UserType } from '@stanfordbdhg/engagehf-models'
 import { createFileRoute, notFound, useRouter } from '@tanstack/react-router'
 import { Users } from 'lucide-react'
 import { queryClient } from '@/app/ReactQueryClientProvider'
 import { callables, docRefs, ensureType } from '@/modules/firebase/guards'
-import { getDocDataOrThrow, UserType } from '@/modules/firebase/utils'
+import { getDocDataOrThrow } from '@/modules/firebase/utils'
 import {
   getUserData,
   userOrganizationQueryOptions,
